+++ src/main/kotlin/model/Source.kt
@@ -4,7 +4,7 @@ data class Source(
     val codec: String? = null,
     val label: String? = null,
     val res_id: Int? = null,
-    val size: Int? = null,
+    val size: Long? = null,
     val status: Boolean? = null,
     val type: String? = null
 )
